(maint) Clean up recap script
Took a shortcut and let claude write the recap script in a pinch, and
I noticed later the complete AI slop it created.

This commit cleans up that recap script to at least be slightly more
bareable.  All of the inline html code has been pulled out to template
files.  More refactoring needs to be done, but this is a start.